Make SplittingStrategy::Semantic require an encoder
Previously if you didn't specify an encoder separately you'd get a runtime exception. It's now required at compile time instead.

I was originally hoping to also set the default to JINA, but Rust doesn't support default values for fields, and it didn't seem like a good idea to set the default SplittingStrategy to Semantic in general.